+++ kernel/platform/pc/include/platform/pic.h
@@ -6,5 +6,5 @@
 
 #pragma once
 
-void pic_map(uint32_t pic1, uint32_t pic2);
+void pic_map(uint8_t pic1, uint8_t pic2);
 void pic_disable(void);
